Minor cleanups (no binary changes)

+++ timecircuits-A10001986/tc_keypad.cpp
@@ -203,10 +203,10 @@ void keypad_setup()
     ettKey.attachPress(ettKeyPressed);
     ettKey.attachLongPressStart(ettKeyHeld);
 
-    ettDelay = (int)atoi(settings.ettDelay);
+    ettDelay = atoi(settings.ettDelay);
     if(ettDelay > ETT_MAX_DEL) ettDelay = ETT_MAX_DEL;
 
-    ettLong = ((int)atoi(settings.ettLong) > 0);
+    ettLong = (atoi(settings.ettLong) > 0);
 #endif
 
     dateBuffer[0] = '\0';
